Windows updates -- accomodating build differences in switching to AL_ and AL_ prefixes; also updated .NET solution and project files -- they may have been out of date
git-svn-id: svn://connect.creativelabs.com/trunk@640 109fce75-3012-0410-ac6a-932efec96c1b

+++ demos/altest/common/altest.cpp
@@ -2278,7 +2278,7 @@ This test alters the frequency of a playing source.
 ALvoid SA_Frequency(ALvoid)
 {	
 	ALuint testSources[2];
-	float root12 = (float)(pow (2, 1/12.0));
+	float root12 = (float)(pow ((float)2, (float)(1/12.0)));
 	float increments[15] = { -12, -10, -8, -7, -5, -3, -1, 0,
 				 2, 4, 5, 7, 9, 11, 12 };
 	int i;
